Refactored packaging of patient controller, removed empty search methods, and client classes that we don't yet need
