+++ argonaut/spotbugs-excludes.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<FindBugsFilter>
+  <!-- Exclusions can be defined here and should include a comment on why the finding can be ignored -->
+  <!-- Example:
+  <Match>
+    <Class name="gov.va.dvp.SomeClass$SomeInnerClass" />
+    <Bug pattern="SIC_INNER_SHOULD_BE_STATIC"/>
+  </Match>
+  -->
+</FindBugsFilter>
