--- argonaut/src/main/java/gov/va/api/health/argonaut/service/mranderson/Profile.java
@@ -1,7 +1,0 @@
-package gov.va.api.health.argonaut.service.mranderson;
-
-public enum Profile {
-  ARGONAUT,
-  DSTU2,
-  STU3
-}
